merged down code from UI money counting branch
set the pause menu back up in a scene called '20171002 - merged stuff'. rebuilt the menu object in the canvas, added the 'PauseGame' script to the Game Mannger, and hooked the menu object into it. Added a text called 'Dosh' to the UI canvas, and added the MoneyUI script. In the PlayerStats.cs script, changed wallet to be a static int so the varriable could be accessed by the MoneyUi.cs script.